Intermediate commit details below
We are trying to expand deck to use the files in build/assets/cards. We
think that the require will help us get there. We don't actually know.
We want to use jasmine to see if we are on the right track. But we are
also exploring how to get jasmine working. Transferring this commit over
so the environment that we know runs jasmine can test.